requirements.txt
requests>=2.24.0
dbus-python>=1.2.0
libtmux>=0.8.0
